reservation_service/requirements.txt
flask==2.2.5
flask_sqlalchemy==2.5.1
psycopg2-binary==2.9.9
kafka-python==2.0.2
sqlalchemy==1.4.46
